Create the docs for willitsync command line utils
Add click-log requirement
Add conda environment file

Add flask-testing to requirements.  It is needed for python-flask,
not for python-client.

Move the conda env file to common location in root.  It isn't necessary
to keep a different environment.yaml file for each of python-client
and python-flask.    A single common file will do.